修复字节数组转float 返回的值却是double类型的bug

+++ hutool-core/src/main/java/cn/hutool/core/util/ByteUtil.java
@@ -336,7 +336,7 @@ public class ByteUtil {
 	 * @return float值
 	 * @since 5.7.18
 	 */
-	public static double bytesToFloat(byte[] bytes) {
+	public static float bytesToFloat(byte[] bytes) {
 		return bytesToFloat(bytes, DEFAULT_ORDER);
 	}
